ai-engine/requirements.txt
fastapi==0.133.0
pydantic==2.12.5
uvicorn==0.41.0
python-dotenv==1.2.1
google-generativeai==0.8.6
groq==1.0.0
Pillow==12.1.1
PyMuPDF==1.27.1
requests==2.32.5
python-multipart==0.0.20